requirements.txt
# Minimal requirements for Python 3.10
python-dotenv
pydantic
requests
openai